doc_processor/requirements.txt
# --- Core Web Framework ---
# Flask is a lightweight web framework for Python used to build the web application interface.
Flask>=2.3.0

# --- Configuration ---
# python-dotenv is used to load environment variables from a .env file for configuration.
python-dotenv>=1.0.0

# --- OCR (Optical Character Recognition) ---
# easyocr is the primary, user-friendly OCR library for text extraction.
easyocr>=1.7.0
# pytesseract is used as a secondary OCR tool, specifically for its Orientation and Script Detection (OSD) capabilities.
pytesseract>=0.3.10
# Pillow is a powerful image processing library, required by both Pytesseract and EasyOCR.
Pillow>=9.5.0
# opencv-python-headless is a dependency for EasyOCR, providing computer vision algorithms without GUI components.
opencv-python-headless>=4.8.0

# --- PDF and Image Handling ---
# pdf2image is a wrapper for pdftoppm to convert PDF files into a sequence of PIL Image objects.
# NOTE: Pinned to 1.17.0 due to environment index lacking >=3.x at install time.
pdf2image==1.17.0
# PyMuPDF (fitz) is used for PDF text extraction and manipulation, providing fast PDF processing capabilities.
PyMuPDF>=1.23.0
# numpy is a fundamental package for scientific computing, often used for handling image data as arrays.
numpy>=1.24.0

# --- Database ---
# SQLAlchemy is a SQL toolkit and Object-Relational Mapper (ORM) that gives application developers the full power and flexibility of SQL.
# In this project, it's used for defining the database schema.
SQLAlchemy>=2.0.0

# --- API Communication ---
# requests is a simple, yet elegant, HTTP library for making requests to APIs, used here to communicate with the Ollama LLM.
requests>=2.31.0
# ollama is the official Python client library for communicating with Ollama LLM servers.
ollama>=0.1.7

# --- Testing ---
# pytest is used for running automated tests
pytest>=7.4.0
pytest-timeout>=2.2.0

# --- Development/Optional Dependencies ---
# These are additional packages that may be useful for development but not strictly required for production

# For advanced logging and debugging (optional)
# colorlog>=6.7.0

# For data validation and type checking (optional) 
# pydantic>=2.0.0